refactor(ui): split FileUploadZone out of ConfigDiffPage
ConfigDiffPage was 520 lines, mostly the page-level diff/merge flow
plus an inline 145-line FileUploadZone subcomponent. The upload zone
has no shared state with the page and was a natural extraction.

Layout:
- ConfigDiffPage.tsx (376)                 — page-level diff + merge flow
- pages/config-diff/FileUploadZone.tsx (169) — drag-and-drop YAML dropzone
                                              + UploadedFile interface

Pure code move; behaviour unchanged. biome / tsc / vitest /
make build all green.
